demo_query/requirements.txt
matplotlib
opencv-python
open3d
torchvision 
tyro
imutils
open_clip_torch